Dictionary for translating model variable names

A dataset containing the model variable names and all information required for obtaining the data.

Usage

dict

Format

A tibble with 40 rows and 14 variables:

model_varname

Variable name in the model equations, must be unique

full_name

Full name/description of the variable

database

Name of the database. Internally implemented are "eurostat" and "edgar"

variable_code

Identifier of the variable if applicable, e.g. Eurostat variable code

dataset_id

Identifier of the dataset where the variable is available, e.g. Eurostat dataset code or link to file on the web

var_col

Name of variable/column in dataset

freq

Frequency of the variable in the dataset. "m" for monthly, "q" for quarterly

geo

ISO 3166-1 alpha-2 country code

unit

Eurostat unit in which the variable is measured, e.g. to choose between different measurements for the same variable

s_adj

Eurostat seasonal adjustment or not

nace_r2

Eurostat identifier for NACE Rev. 2 classification

ipcc_sector

EDGAR IPCC National Greenhouse Gas Inventories, see https://www.ipcc-nggip.iges.or.jp/public/2006gl/index.html. "TOTAL" is not an official IPCC code but is internally interpreted to use country totals.

cpa2_1

Eurostat identifier for Classification of Products by Activity (CPA)

siec

Standard International Energy Product Classification, e.g. for Eurostat

...

Additional filters across the database. Need to have exactly the same column name as in the respective database

Source

Own compilation, codes from https://ec.europa.eu/eurostat and https://edgar.jrc.ec.europa.eu/dataset_ghg70