Estimate the specific module using indicator saturation

Estimate the specific module using indicator saturation

Usage

estimate_module(
  clean_data,
  dep_var_basename = "imports_of_goods_and_services",
  x_vars_basename = c("gross_capital_formation",
    "household_and_npish_final_consumption_expenditure"),
  use_logs = c("both", "y", "x"),
  trend = TRUE,
  ardl_or_ecm = "ardl",
  max.lag = 4,
  saturation = c("IIS", "SIS"),
  saturation.tpval = 0.01,
  max.block.size = 20,
  gets_selection = TRUE,
  selection.tpval = 0.01
)

Arguments

clean_data

An input data.frame or tibble. Must be the output of clean_data() to fit all requirements.

dep_var_basename

A character string of the name of the dependent variable as contained in clean_data() in a level form (i.e. no ln or D in front of the name).

x_vars_basename

A character vector of the name(s) of the independent variable(s) as contained in clean_data() in a level form (i.e. no ln or D in front of the name).

use_logs

To decide whether to log any variables. Must be one of "both", "y", or "x". Default is "both".

trend

Logical. To determine whether a trend should be added. Default is TRUE.

ardl_or_ecm

Either 'ardl' or 'ecm' to determine whether to estimate the model as an Autoregressive Distributed Lag Function (ardl) or as an Equilibrium Correction Model (ecm).

max.lag

The maximum number of lags to use for both the AR terms as well as for the independent variables.

saturation

Carry out Indicator Saturation using the 'isat' function in the 'gets' package. Needes is a character vector or string. Default is 'c("IIS","SIS")' to carry out Impulse Indicator Saturation and Step Indicator Saturation. Other possible values are 'NULL' to disable or 'TIS' or Trend Indicator Saturation. When disabled, estimation will be carried out using the 'arx' function from the 'gets' package.

saturation.tpval

The target p-value of the saturation methods (e.g. SIS and IIS, see the 'isat' function in the 'gets' package). Default is 0.01.

max.block.size

Integer. Maximum size of block of variables to be selected over, default = 20.

gets_selection

Logical. Whether general-to-specific selection using the 'getsm' function from the 'gets' package should be done on the final saturation model. Default is TRUE.

selection.tpval

Numeric. The target p-value of the model selection methods (i.e. general-to-specific modelling, see the 'getsm' function in the 'gets' package). Default is 0.01.

Value

A list containing all estimated models, with the model with the smallest BIC under 'best_model'.

Examples

sample_data <- dplyr::tibble(
  time = rep(seq.Date(
    from = as.Date("2000-01-01"),
    to = as.Date("2000-12-31"), by = 1
  ), each = 2),
  na_item = rep(c("yvar", "xvar"), 366), values = rnorm(366 * 2, mean = 100)
)
sample_data_clean <- aggregate.model:::clean_data(sample_data, max.lag = 4)
aggregate.model:::estimate_module(sample_data_clean, "yvar", "xvar")
#> $isat_list
#> # A tibble: 5 × 3
#>      ar    BIC isat_object
#>   <int>  <dbl> <list>     
#> 1     0 -2325. <isat>     
#> 2     1 -2312. <isat>     
#> 3     2 -2297. <isat>     
#> 4     3 -2281. <isat>     
#> 5     4 -2254. <isat>     
#> 
#> $best_model
#> 
#> Date: Sat Sep  9 12:13:24 2023 
#> Dependent var.: y 
#> Method: Ordinary Least Squares (OLS)
#> Variance-Covariance: Ordinary 
#> No. of observations (mean eq.): 366 
#> Sample: 1 to 366 
#> 
#> SPECIFIC mean equation:
#> 
#>               coef   std.error    t-stat   p-value    
#> mconst  4.60567841  0.00069286 6647.3772 < 2.2e-16 ***
#> q_3    -0.00530316  0.00161351   -3.2867 0.0011144 ** 
#> iis63  -0.02628558  0.00924386   -2.8436 0.0047186 ** 
#> iis114 -0.03043077  0.00924386   -3.2920 0.0010944 ** 
#> iis134 -0.02721326  0.00924386   -2.9439 0.0034535 ** 
#> iis290 -0.03403534  0.00930863   -3.6563 0.0002944 ***
#> sis181  0.01271652  0.00339057    3.7506 0.0002059 ***
#> sis190 -0.01133705  0.00318255   -3.5622 0.0004177 ***
#> sis325 -0.01311443  0.00319038   -4.1106 4.903e-05 ***
#> sis335  0.01232284  0.00333949    3.6900 0.0002593 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Diagnostics and fit:
#> 
#>                   Chi-sq df p-value  
#> Ljung-Box AR(1)   4.1411  1 0.04185 *
#> Ljung-Box ARCH(1) 1.3619  1 0.24321  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>                            
#> SE of regression    0.00922
#> R-squared           0.17635
#> Log-lik.(n=366)  1200.96878
#> 
#> $args
#> $args$clean_data
#> # A tibble: 366 × 28
#>    index time       trend  yvar  xvar ln.yvar ln.xvar D.ln.yvar D.ln.xvar
#>    <int> <date>     <dbl> <dbl> <dbl>   <dbl>   <dbl>     <dbl>     <dbl>
#>  1     1 2000-01-01     1 100.  100.     4.61    4.61  NA        NA      
#>  2     2 2000-01-02     2 100.   99.2    4.61    4.60   0.00118  -0.00903
#>  3     3 2000-01-03     3  99.7 100.     4.60    4.61  -0.00459   0.00939
#>  4     4 2000-01-04     4 100.  101.     4.60    4.61   0.00250   0.00651
#>  5     5 2000-01-05     5 101.   99.3    4.61    4.60   0.00659  -0.0142 
#>  6     6 2000-01-06     6 100.  100.     4.61    4.61  -0.00438   0.0115 
#>  7     7 2000-01-07     7 101.  101.     4.61    4.61   0.00568   0.00213
#>  8     8 2000-01-08     8 101.   99.1    4.61    4.60  -0.00154  -0.0158 
#>  9     9 2000-01-09     9  99.7 102.     4.60    4.62  -0.00863   0.0243 
#> 10    10 2000-01-10    10 100.  102.     4.61    4.63   0.00534   0.00724
#> # ℹ 356 more rows
#> # ℹ 19 more variables: L1.D.ln.yvar <dbl>, L1.D.ln.xvar <dbl>,
#> #   L1.ln.yvar <dbl>, L1.ln.xvar <dbl>, L2.D.ln.yvar <dbl>, L2.D.ln.xvar <dbl>,
#> #   L2.ln.yvar <dbl>, L2.ln.xvar <dbl>, L3.D.ln.yvar <dbl>, L3.D.ln.xvar <dbl>,
#> #   L3.ln.yvar <dbl>, L3.ln.xvar <dbl>, L4.D.ln.yvar <dbl>, L4.D.ln.xvar <dbl>,
#> #   L4.ln.yvar <dbl>, L4.ln.xvar <dbl>, q_2 <int>, q_3 <int>, q_4 <int>
#> 
#> $args$dep_var_basename
#> [1] "yvar"
#> 
#> $args$x_vars_basename
#> [1] "xvar"
#> 
#> $args$use_logs
#> [1] "both"
#> 
#> $args$ardl_or_ecm
#> [1] "ardl"
#> 
#> $args$max.lag
#> [1] "ardl"
#> 
#>